refactor: form and form inputs
-Move form input data from data file into contact form component to improve readability. -Create a folder to manage forms. -Move form inputs, contact form and all related components into form folder.
